double for all types of operations

+++ app/src/main/java/com/app/calculator/Model/CalculatorLogic.java
@@ -5,13 +5,13 @@ public class CalculatorLogic {
     public double addition (double a, double b){
         return a + b;
     }
-    public int subtract (int a, int b){
+    public double subtract (double a, double b){
         return a - b;
     }
-    public int multiply (int a, int b){
+    public double multiply (double a, double b){
         return a * b;
     }
-    public int divide (int a, int b){
+    public double divide (double a, double b){
         return a / b;
     }
 }
